Ground/landing hitbox for piledrive
